Move existing lint into shared file
