+++ src/lintlist/mod.rs
@@ -65,7 +65,7 @@ vec![
         group: "pedantic",
         desc: "Inside an async function, holding a MutexGuard while calling await",
         deprecation: None,
-        module: "await_holding_lock",
+        module: "await_holding_invalid",
     },
     Lint {
         name: "await_holding_refcell_ref",
