+++ tests/ui/await_holding_invalid.rs
